refactor: 패키지분리
- API 응답과 뷰 응답을 분리해줌
