+++ src/main/java/roomescape/controller/GlobalExceptionHandler.java
@@ -1,4 +1,4 @@
-package roomescape.controller.reservation;
+package roomescape.controller;
 
 import org.springframework.http.HttpStatus;
 import org.springframework.web.bind.annotation.ExceptionHandler;
