+++ src/main/java/roomescape/controller/web/admin/AdminThemePageController.java
@@ -1,15 +1,15 @@
-package roomescape.controller.user;
+package roomescape.controller.web.admin;
 
 import org.springframework.stereotype.Controller;
 import org.springframework.web.bind.annotation.GetMapping;
 import org.springframework.web.bind.annotation.RequestMapping;
 
 @Controller
-@RequestMapping("/reservation")
-public class ReservationPageController {
+@RequestMapping("/admin/theme")
+public class AdminThemePageController {
 
     @GetMapping
-    public String getUserPage() {
-        return "reservation";
+    public String getThemePage() {
+        return "/admin/theme";
     }
 }
