+++ src/test/java/roomescape/controller/api/reservation/ThemeControllerTest.java
@@ -1,4 +1,4 @@
-package roomescape.controller.reservation;
+package roomescape.controller.api.reservation;
 
 import io.restassured.RestAssured;
 import io.restassured.http.ContentType;
