+++ src/test/java/roomescape/controller/web/ReservationPageControllerTest.java
@@ -1,4 +1,4 @@
-package roomescape.controller.admin;
+package roomescape.controller.web;
 
 import io.restassured.RestAssured;
 import org.junit.jupiter.api.DisplayName;
@@ -7,14 +7,13 @@ import org.springframework.boot.test.context.SpringBootTest;
 import org.springframework.test.annotation.DirtiesContext;
 
 @SpringBootTest(webEnvironment = SpringBootTest.WebEnvironment.DEFINED_PORT)
-@DirtiesContext(classMode = DirtiesContext.ClassMode.BEFORE_EACH_TEST_METHOD)
-public class AdminThemePageControllerTest {
+class ReservationPageControllerTest {
 
     @Test
-    @DisplayName("/admin/theme 페이지를 응답한다.")
-    void responseThemePage() {
+    @DisplayName("사용자의 예약 페이지를 응답한다.")
+    void getUserPage() {
         RestAssured.given().log().all()
-                .when().get("/admin/theme")
+                .when().get("/reservation")
                 .then().log().all()
                 .statusCode(200);
     }
