+++ src/test/java/roomescape/controller/web/admin/AdminThemePageControllerTest.java
@@ -1,4 +1,4 @@
-package roomescape.controller.admin;
+package roomescape.controller.web.admin;
 
 import io.restassured.RestAssured;
 import org.junit.jupiter.api.DisplayName;
@@ -8,14 +8,16 @@ import org.springframework.test.annotation.DirtiesContext;
 
 @SpringBootTest(webEnvironment = SpringBootTest.WebEnvironment.DEFINED_PORT)
 @DirtiesContext(classMode = DirtiesContext.ClassMode.BEFORE_EACH_TEST_METHOD)
-class AdminReservationTimePageControllerTest {
+public class AdminThemePageControllerTest {
 
     @Test
-    @DisplayName("/admin/time 페이지를 응답한다.")
-    void responseAdminPage() {
+    @DisplayName("/admin/theme 페이지를 응답한다.")
+    void responseThemePage() {
         RestAssured.given().log().all()
-                .when().get("/admin/time")
+                .when().get("/admin/theme")
                 .then().log().all()
                 .statusCode(200);
     }
+
+
 }
